+++ src/include/parser/nodes/NodeStruct.hpp
@@ -17,7 +17,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 extern Type* checkForTemplated(Type* type);
 
 struct StructPredefined {
-    int element;
+    size_t element;
     Node* value;
     bool isStruct;
     std::string name;
